ODROID: uboot: Change uboot.img including condition.
- Changed uboot.img including condition. The default setting is not
  including the uboot.img when user didn't build it.

+++ selfinstall/selfinstall.sh
@@ -9,7 +9,9 @@ TARGET_IMAGE=$TARGET_PATH/selfinstall.img
 HOST_OUT=`get_build_var HOST_OUT_EXECUTABLES`
 SGDISK_HOST=$HOST_OUT/sgdisk
 
-dd if=$TARGET_PATH/uboot.img of=$TARGET_IMAGE bs=512 seek=49152
+if [ -f $TARGET_PATH/uboot.img ]; then;
+	dd if=$TARGET_PATH/uboot.img of=$TARGET_IMAGE bs=512 seek=49152
+fi
 
 dd if=$TARGET_PATH/fat.img of=$TARGET_IMAGE bs=512 seek=2048
 dd if=$TARGET_PATH/misc.img of=$TARGET_IMAGE bs=512 seek=53248
